Add a custom action to find a mapping for a site
In order to identify the mapping for a hit, we need to be able to
find a mapping given an uncanonicalized path for a site. We've done
it through a custom action which redirects to the mapping so that
we don't have to canonicalize all the paths on a hits page and make
lots of queries to try to find the mappings for them all.

If a mapping is found for the canonicalized path, a redirect to it
is returned. If no mapping is found, a redirect to the 'new mapping'
page is returned, with the canonicalized path shown in the form.
If the path canonicalizes to a homepage (e.g. '/') then it redirects
back to the previous page if possible (or to the site's mapping
index if not) with a notice that it's not possible to edit homepage
mappings.

The route match allows us to refer to the path in its singular
form - 'find_site_mapping' instead of 'find_site_mappings' which
it otherwise would be called because it is nested under the mappings
collection. We're only finding one mapping for one given path on
the site, so this is a more appropriate name.