Change gds_transition_manager to admin
"gds_transition_manager" didn't describe everyone in the group. Crudely it will
probably be anyone at GDS. Potentially, we could set everyone at GDS to be a
member of the GDS organisation in Signon, and then have GDS be a special case
in Transition, but because Whitehall doesn't have GDS as an organisation, we
can't really do this.