Add 'managed_by_transition' flag to Site
This flag defaults to true so that new sites added to the Transition
app first will be automatically managed there. The flag is set to
false by the migration for all existing sites, because at the moment
all sites are still managed by Redirector.

The flag represents the intended state of the world; requests will
not actually be sent to Redirector or Bouncer until the CDN has been
configured for each host, according to its site's flag. A name for
this column indicating that it reflects intention rather than reality
would be inconveniently cumbersome, and hopefully the delay between
changing the flag and the CDN configuration being updated will be
small in practice.