Update access control to allow for multiple parent orgs
A user can edit an organisation's sites if they are a member of one
of the organisation's parent organisations. A user can be a member of
only one organisation in Signon, so it is not possible for a user to
be a member of all of an organisation's parents if there are several.