Fetch Whitehall orgs from production in development
We frequently need to run the rake import:all task in development,
which currently requires Whitehall to be running locally, with
up-to-date data. This change means that the task will call the
production organisations API from dev instead, which avoids the
overhead of keeping Whitehall up to date and running locally.