If some updates fail, display only those again
This is difficult to test, because it would involve creating an
invalid mapping to attempt to update, but it's a real edge case.
The only way I can think of for some mappings to not be updateable
is if we accidentally import a homepage mapping (bypassing validation
in the import) which is then invalid because it doesn't have a path,
and so would not be saved with the new data.