Use a gem to generate breadcrumbs
The old breadcrumb was going to be difficult to change to work with a proposed
Site#show page between Organisation#show and Mappings#index/Hits#index.

Using the gem, we've got less code in the app, although we have gained a single
dependency.

It was also becoming a little hard to read.