Store transition_status on a site, including 'indeterminate'
Apart from when we occasionally update a site's status to
'indeterminate', the status will only change when DNS details are
imported, so it makes sense to store the status for all sites and
update it in a later import task. This avoids having to query all a
site's hosts to determine its status when loading the organisation
show page, and keeps the status override logic in one place.

This incorporates the styling for indeterminate status from the
prototype ( dbc17f4 ), with minor updates for Bootstrap 3.