Set whodunnit to user.name and set user_id
whodunnit is the user id as a string by default in Paper Trail, but
we set it to the user's name when versions are created through the
mappings controller and also set the version.user_id field. We need
to do this whenever versions are created outside the controller too
for consistency.