Update version data in mapping model spec
This is the obvious place to look for what a version's whodunnit
and user_id fields should be, so it was confusing to have tests
using PaperTrail's default data rather than what we are saving
from changes made via the controller. When we edit mappings in
imports or in any other way which does not use the mappings
controller, we need to make sure that the versioning data is
created consistently, so it's helpful to have that documented
accurately here.