Reduce the amount of data put in the session
We store a 'return_url' for the mappings page for each site. This enables us to
return the user to where they were before eg bulk editing.

The problem with this is that cookies are limited in size. When we exceed that
size, nginx rejects the response because the header is too large, and returns a
502.

By storing the fullpath, rather than the URL, we reduce the problem. Whilst
there are options for solving it, we need to be careful about the security
implications.