Don't use site_return_path for bulk adding
It makes more sense to just return to the site mappings index after
creating multiple mappings, because the user is unlikely to have added
mappings which would then all be visible in the filtered and paginated
view. The main mappings index showing the total number of mappings will
probably be more helpful for the user to check that mappings were
correctly created if they want to.