Pass return URL through bulk edit forms
This no longer stores a return URL per site in the session cookie,
which resolves the bug involving the session cookie growing too large
for nginx to accept the request. It also means that if a user is
working on a single site across multiple tabs, their histories will
not get confused.