Give Whitehall more time to finish mappings export
We're currently sometimes downloading the file before it has finished being
generated. This means we only get the file as it stands when we download it.

For UKBA, which is near the end of the file, this meant we weren't getting any
of their mappings.

It currently takes ~42 minutes to generate the file; we were allowing 30 mins.
Let's give it 1hr 15mins.