Don't encode < & >, but use URI::BLURI's functionality to handle them
Following a conversation with Jenny and Russell, it has been decided
to not encode < and >, but to use the functionality of the Addressable
gem included in BLURI (in optic14n) to handle them. This makes it
easier for the user to see what their paths are on the confirmation
screen without having to understand encodings.

Also edit the tests so that the expected output in this case doesn't
contain encodings.
Handle the exception raised by just "<" or ">" being input into the
bulk add input field by returning an understandable error to the user
about the path being invalid.