Override archived mappings from Whitehall
We want users to be able to bulk-add mappings as 410s, and then
have them automatically updated later to redirect when their content
is published in Whitehall. An existing mapping will now not be
updated from the Whitehall mappings CSV only if its last edit was
made by a human and it is a redirect.