Check the site abbr when showing a mapping's history
The versions controller was only using the mapping id and ignoring
the site_id param, so it was possible to see a mapping's history
even when the site abbr in the URL did not match the mapping's site's
abbr. Now it 404s in this case. (The page always showed the correct
site's details in the breadcrumb and elsewhere.)