Fix error in mapping saved modal on non-site pages
It is possible to have saved_mapping_ids in your session, but to be viewing a
page which doesn't relate to a site, so @site isn't set. Before now, this would
mean seeing a 5xx error page.

I was able to reproduce this by saving mappings in one tab and quickly
refreshing the homepage in another. There are lots of other ways to trigger the
problem.