Fix lookup of originating user in edited_by_human?
This method looks for the last user recorded as having edited the
mapping, and tries to tell if the user is human or a robot. The
'originator' is an alias for the version's 'whodunnit', and we store
the user's name in this field rather than the user id (which PaperTrail
defaults to). We need to try to find the user by name rather than id,
otherwise this method will never return true in real use.

The tests for this method were passing before because in model specs
the PaperTrail config was not being correctly set, so it was saving
the user's id in the whodunnit field rather than the name, so find_by_id
was able to find the user.

Controller specs are unaffected by this change, because PaperTrail
prefers to have its config on the ApplicationController and will
always use that configuration when we go through the controller. In
controller specs we have always set whodunnit to the name, so in that
context our tests are still valid.

- Add as_user to the mapping factory, so that we can specify a user
  by whom the mapping should appear to be created. The user defaults
  to a stub user if it is not given, none is already set and
  PaperTrail is enabled.

- Create a Transition::History module, so that we have somewhere to
  put methods which will make it easier to do the right thing in future