Ensure that PaperTrail config is set before saving a mapping
If PaperTrail is enabled, we should make sure that we configure it
consistently wherever it is used. This will make it much easier to
use it correctly in future, and clearly sets out what we expect to
be set.

We have so far been using the correct configuration everywhere in
our code, but often not in tests, especially non-controller specs or
anywhere which requires mappings to be created as setup with
versioning: true.