Disable Sidekiq retries
Sidekiq will perform 25 retries over approximately 21 days using an exponential
backoff algorithm. We probably don't want mappings being changed days afterward
so let's never retry.

This is fairly drastic - we could have 6 retries which would be attempted over
6 minutes, but the MappingsBatch state field behaviour is incompatible with
retries - it would mark them as failed after the first attempt.

If we wanted to add retries back in, we'd need to change MappingsBatch's model
of state to cope with it.