Process small batches inline
This makes much more sense when adding (possibly updating) a single mapping or
a few mappings, which is definitely the normal case.

This reinstates the modal confirmation dialogue for small batches.

The specs for mappings_batch_outcome_presenter.rb were cribbed from the specs
for BulkAdder.

Reverted features changed earlier in this branch which run small batches to
look for the modal, rather than the "x of y processed" flash message used for
large batches.

20 was chosen as the upper limit for a "small" batch because it should be easily
processable within the 15 second timeout we have for admin apps, but is just
above the average batch size according to Google Analytics.