Updating query_params causes refresh
When a site's significant query_params have changed, what a given Hit or
HostPath canonicalises to can change. This means that we might:
* show the wrong hits count for a mapping
* link to the wrong mapping from a hit
* not create new mappings from host_paths

Until now, we've ignored this problem. This change causes the relationship
between Mappings and HostPaths to be cleared and reconnected according to the
new canonicalisation rules. In turn, the connection to Hits is updated.