Add a 'type' column to the three mappings tables
This column stores the mapping type as a word string rather than as
an HTTP status code string. There isn't an appropriate HTTP status code
to use for unresolved mappings, so this is the cleanest way to store
users' decisions about mappings in the long run.

The new column has the same null constraints as the existing
http_status column on that table in each case.

This commit also adds validations for Mapping.type and sets the type
from http_status before validation. The type is not yet being set
correctly everywhere, but all the existing tests pass.