Make the mapping edit feature pass again
The specs for the ToggleMappingFormFields js module were not being
run because the filename didn't end 'spec.js', and after renaming
the file I found that 4 of those tests were failing (on master as
well as on this branch). I've updated the tests so that they look
ore up-to-date, but they are still failing and the output from
rake jasmine:ci isn't very helpful.