Set global_type in site yaml import
We need to import sites which are still handled by redirector, so
can't switch to using redirect/archive as global values in all site
yaml files, so should convert 301/410 to redirect/archive on import.

Before this, global_http_status was being set on import but not
global_type, so any changes to global statuses would not have been
reflected in the global_type column. This wouldn't have been a problem
for Bouncer, which isn't yet using the new column, but various things
in Transition are using the new column already so for example users
wouldn't see the message on the site dashboard about the global type/status.

The import now only sets global_type explicitly; the global_http_status
is set for now by the before_validation callback.