Increase default limit of tags to show to 50
MOJ are using more than 10 tags, and since we limit the number of
tags which are shown in the dashboard and in the filter dropdown,
they had no way of getting back to the mappings tagged with the less
frequent tags. As a short-term measure we can increase the default
limit to 50, which shouldn't be unmanageable and it's likely that very
few sites have anything like this many tags anyway. As a more robust
solution we should make it possible to type in the tag filter dropdown
to find any tag, but that is a separate story.