Set global_http_status from global_type even if it is already set
Now that the site yaml import sets global_type only, it is the source
of truth, and so if a global status is removed from the site yaml the
change of global_type to nil should also set global_http_status to nil,
rather than leaving it as the original value. If it had been a redirect,
the global_new_url would have been removed by the import so the redirect
would error with the current behaviour of Bouncer (still using the old
column).