Only validate New URL in whitelist for redirects
Other validations of New URL are only done if the type is a redirect. This
means that if you typed in something invalid for New URL and then changed the
type to Archive, it won't reject the change you actually wanted and complain
about the value of a field which we have hidden from view with JavaScript.