Refactor message for a hit with an unresolved mapping
We want to show this message for any hit with an unresolved mapping,
regardless of the hit's http_status. Having a separate helper method
for this case makes it clearer that this is the behaviour we expect,
and that we're not relying on any comparison between the hit's http_status
and its mapping's type to determine what to show here. It also means
that this message is shown by one code path, rather than it coming from
different places depending on the hit-mapping comparison.

This also moves the check for the hit having a mapping back into the
views, which makes it clearer there that if the hit has no mapping then
nothing extra will be displayed.