Add an import preview page
This displays counts of each type of mapping to be created, and a
table which currently contains all of the batch's entries (rather
than the first 20).

The table duplicates small parts of the _mappings_table partial, but
since that partial includes many behaviours we don't need here it is
simpler to have this duplication instead of trying to make the partial
usable in this context as well.

This also adds a feature for the happy path (at the moment it only goes
as far as the preview page).