Validate all new URLs in the batch for length
To do this we have added a class, EachInCollectionValidator, which
calls a validator class on each item in the collection. It puts the
errors on the given attribute rather than each value individually,
so that they will attach to the correct field in the form.

We are doing this because we want to be able to use activerecord-import
to import the records in the batch without going through validations
on save.