Move header checking onto ImportBatchRow
This simplifies the detection of header rows and other rows which can't
be processed and makes it more robust by handling leading whitespace.

ImportBatchRow is responsible for taking a parsed row of CSV and
presenting it so that it can be validated as part of the batch and
then have an entry created from it. It shouldn't be the responsibility
of deduplicated_csv_rows to know enough about the CSV row to pass it
as old and new values.