Validate that abbr is slug-safe
We currently have one which is `directgov_www.actonco2`. This is a problem
because Rails interprets the . as indicating a format, but it is a format it
doesn't support, so Rails returns a 402.

An alternative fix would be to tweak the Site#to_param to make abbr slug-safe,
but doing it this way means that the abbr is always the slug and vice versa.

The mappings_csv_spec had to be changed to use a valid abbr.