Move MappingsProgress into a controller concerns directory
This is the idiomatic place for concerns to live (especially as of
Rails 4) so seems sensible, even though we're not using ActiveSupport::Concern
which would generally be used by concerns which live here.

Having this code in app/ rather than lib/ also means that the dev server
reloads it for us.