Add DisableColumns as a concern
Functionality is the same as it was before, which isn't perfect...

This currently only works when instantiating a new object; fetching
one from the database and then trying to access the disabled columns
doesn't raise an error.

The error raised by ImportBatch.new.type is actually
ActiveModel::MissingAttributeError and not NoMethodError - the test
asserting that NoMethodError should be raised was still passing,
presumably because MissingAttributeError inherits from NoMethodError,
but I've made the test more specific for accuracy.

However, ImportBatch.new.foo *does* raise NoMethodError, so it is
still treating the disabled columns differently from ones which
don't exist in the database.