Add titles and furls to sites
Sites can already have homepages which differ from their owning organisation's
homepage. Given that the furl attribute is simply a shorter form of the full
URL for a site, it follows that they have furls too.

We are currently inconsistent with how we apply the (new) homepage for a site.
  * If you look at the site in Transition, it says the new homepage is the one
    specified for that site in Redirector.
  * If you visit the old homepage, you get redirected to the site's new
    homepage, which is correct.
  * However, if you visit a page with an archive/unresolved mapping, it links
    to the organisation's new homepage, which is unhelpful or even
    inappropriate.

Therefore, we should change Bouncer's archive behaviour so that it links to the
site's new homepage. To make it consistent, we then need to make it use the
furl (if any) and a title which describes the site, rather than the
organisation. We need to make these changes to support that.