Connect to the database via localhost in dev and test
Doing this for postgres means that we don't need to add any pg_hba.conf
rules in dev for the app to be able to authenticate as the transition
role, because one of the default rules already allows it. Using host
authentication for postgres will also make dev more similar to production.

This works for mysql in dev too, so we might as well do this on master
to minimise the diff for postgres branches.

This file is overwritten on deployment so this change makes no difference
for production.