Rely on peer auth in development and test environments
It's simpler to have a postgres role with the same name as the Unix
user and use peer authentication in these environments. This is
different from production, but should work for CI as well.

To do this we need to not specify user, password or a host to use
for connecting to the database.