Extend rake db:create to enable pgcrypto extension
This extension can only be enabled by a superuser, but we don't want
to be running the app or tests as a superuser, so we need to do
this as the postgres user.

This allows us to use db:setup and db:reset as usual in dev and test.

This is a temporary measure and can be removed when we no longer need
pgcrypto (after removing path_hash).