Drop and recreate the database each time in CI
db:schema:load isn't able to keep the schema completely as it is
in the schema file, which is particularly a problem for branch
builds. Dropping the database each time is cleaner and ensures that
tests are running with only the specified tables in the database.
Now that we are enabling pgcrypto as part of db:create, db:reset
can do this for us.