Set redis namespace to 'transition-postgres'
This is a temporary measure to ensure that when the postgres-master
branch is deployed to preview it doesn't share a queue in redis for
sidekiq jobs. The sidekiq workers already have different names, but
without this namespacing change they are all picking up jobs from
the same queue, which is possibly causing jobs for transition to
fail. I think they could be being picked up by transition-postgres's
workers, which wouldn't be able to do anything in the postgres
database right now but would leave the jobs in an unfinished state
in the mysql database, so they would appear in the flash message
as though they hadn't started.

We can set the namespacing in the config which is overwritten on
deployment, but should switch to do that on master, rebase this
change onto master and revert this change to minimise differences
between the branches.