Link to mappings CSV export, for admins only
We don't want to encourage use of spreadsheets, but we do occasionally get
asked to provide them. Let's make life easy on ourselves by giving us a means
to provide the spreadsheet easily. It also enables advanced users to search and
filter in ways we don't intend to support (eg look for mappings matching a
regex).