Fix choosing a date from category hits graph pages
If you were viewing the hits in a category, for example:

/sites/businesslink/hits/category?category=errors

The code was overwriting the querystring and so omitting the category
parameter. Following that link would trigger a 5xx error because category was
empty when it shouldn't be.

Using the URI library, we can add or overwrite just the relevant part of the
querystring.