Store path relative to Rails.root when tracking hits files
We use timestamped directories for releases in production. This
means that on every deploy the absolute filenames for hits files
change, even if the hits files in transition-stats and pre-transition-stats
haven't changed. All hits files are therefore re-imported after every
deploy, which is unnecessary and takes a long time.

To fix this, this stores the filename relative to Rails.root, thereby
excluding the part of the path which changes on deploy. @rgarner thought
he had a reason for not doing this originally, but can't remember it now.

It would be good to avoid this in production, but in preview it is
even more important at the moment. Preview is deployed to more frequently
than production, and the overnight data sync from production also
causes timestamp mismatches. We're trying to compare performance of
the hits import between MySQL and Postgres at the moment, and that
is very hard when the import mostly takes much longer than it needs to,
and longer than it would be in production.