Compromise hits indices to balance import performance
Compromise indices to make import/view hits work after discovery of
very bad index selection on 9.1 in Preview (blocking single hit file imports
on a single query taking > 20mins). This is an admission that
we don't know what's gone wrong in Preview but we know how to fix it.
Preview doesn't select the "bad" import index when it isn't there, and hits view
can use these two in concert.

We don't know that 9.1 caused this, because it wasn't happening in dev
on 9.1, so we're assuming an environmental difference. We don't know
whether that environmental difference will affect 9.3.x in the same way.
We should try reverting to the two 3/4-column index solution on 9.3.x
later by reverting or erasing this commit.

This is less than ideal because we know extra indices slows import down.
On the plus side it speeds up the viewing of hits (particularly for
summary/category) by around 100%.