Compromise hits indices to balance 9.1 performance
Compromise indices to make import/view hits work after discovery of
very bad index selection on 9.1 in Preview (blocking single hit file imports
on a single query taking > 20mins). This is an admission that
we don't know what's gone wrong but we know how to fix it (9.1 doesn't
select the "bad" import index when it isn't there, and hits view can use
the two in concert).

It's less than ideal, and hopefully we can go back to the two 3/4-column
index solution for hits if we can use 9.3.x by reverting or erasing this
commit.

This is less than ideal because we know extra indices slows import down.
On the plus side it will speed up the viewing of hits (particularly for
summary/category) by 100-200%.