Run DNS details import hourly only during the working week
This import seems to be able to cause a deadlock if it runs during the
mysql-to-postgres sync, which can cause the sync job to fail. We only
need the DNS details to be updated hourly while users are likely to be
in the app, so only run it during workdays so that it will not be running
early in the morning, potentially conflicting with the sync job.