Stop using pgcrypto extension
Now that we've got rid of path_hash entirely, we don't need this
extension. Even though there's a migration here to drop it, we need
to remove the extension from Puppet and make sure that it is dropped
in all environments by hand before we deploy this. Otherwise the
migration will fail because the transition Postgres role isn't a
superuser. We had a migration to enable this extension, so should
have one to remove it too, even though it isn't much use to us.