Keep the (empty) data directory in the repo
We don't want the Jenkins jobs which run our various import tasks to
have to care about whether or not this directory exists already, so
let's ensure that it always exists after a deploy. We do this for the
log directory already (although that directory is overwritten with a
symlink on deploy).

This .gitignore change keeps the data directory, but ignores everything
in it except for the .gitkeep file, so we won't accidentally commit the
notmodules into this repo.

The presence of the .gitkeep file shouldn't cause any problems for the
imports - they always look for files inside one of the subdirectories,
not in the data directory directly.